+++ ui/mods/com.pa.quitch.qquellerai/new_game.js
@@ -410,7 +410,7 @@ function quellerAIPersonalities() {
     });
   } catch (e) {
     console.error(e);
-    console.error(JSON.stringify(e));
+    console.error("Queller AI: " + (e.stack || e.message || e));
   }
 }
 quellerAIPersonalities();